requirements.txt
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.5
decorator==4.4.2
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
Flask==3.1.3
Flask-APScheduler==1.13.1
Flask-WTF==1.2.2
google-auth==2.48.0
google-genai==1.66.0
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
instagrapi==2.3.0
instaloader==4.15
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
moviepy==1.0.3
numpy==2.4.2
pandas==3.0.1
pillow==12.1.1
proglog==0.1.12
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pycryptodomex==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.48
tenacity==9.1.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
websockets==16.0
Werkzeug==3.1.6
WTForms==3.2.1
